Refactor animations and add LED state persistence
Refactored `BackgroundAnimator` and `IAnimator` to improve immutability, thread safety, and encapsulation. Enhanced `AnimationSettings` by switching to `ushort` for better type safety.

Introduced `BrightnessLevel` class for managing LED brightness with validation and immutability. Added `ILed` interface for LED control abstraction.

Rewrote `LedDiode` to support PWM-controlled brightness with smooth animations, state persistence via `IStateStorage`, and customizable easing styles. Added `LedState` class to persist LED state across power cycles.

Updated project files to include new classes and dependencies. Removed legacy GPIO-based LED control logic. Improved documentation and examples for `LedDiode`, `ILed`, and `IAnimator`.

General improvements include better resource management, thread safety, and code readability.

+++ src/Axtox.IoT.Common/Animations/Settings/AnimationSettings.cs
@@ -53,7 +53,7 @@ namespace Axtox.IoT.Common.Animations.Settings
         /// to complete immediately with the target value.
         /// </para>
         /// </remarks>
-        public int DurationInMilliseconds { get; set; } = 100;
+        public ushort DurationInMilliseconds { get; set; } = 100;
 
         /// <summary>
         /// Gets or sets the update interval in milliseconds between animation frames.
@@ -79,7 +79,7 @@ namespace Axtox.IoT.Common.Animations.Settings
         /// The actual update rate may vary based on system load and thread scheduling.
         /// </para>
         /// </remarks>
-        public int UpdateIntervalInMilliseconds { get; set; } = 5;
+        public ushort UpdateIntervalInMilliseconds { get; set; } = 5;
 
         /// <summary>
         /// Gets or sets the easing function style applied to the animation.
